Create BombNumbers.js
Afunction that receives two parameters: sequence of numbers and a special bomb number with a certain power. 
The task is to detonate every occurrence of the special bomb number and according to its power his neighbors from left and right. Detonations are performed from left to right and all detonated numbers disappear. 
The output is the sum of the remaining elements in the sequence.