Create Cats.js
A function that receives array of strings in the following format '{cat name} {age}'.
A class Cat is created. It receives in the constructor the name and the age parsed from the input and a function named "meow" that will print "{cat name}, age {age} says Meow" on the console.