Create Songs.js
Task is to define a class Song, which holds the following information about songs: typeList, name and time. Input comes as an array.
The first element n is the number of songs. Next n elements are the songs data in the following format: "{typeList}_{name}_{time}", and the the last element will be Type List / "all".
Function prints only the names of the songs which are from that Type List / All songs.