Create Towns.js
The task is to create and print objects from a text table. 
Input comes as an array of strings, where each string represents a table row, with values on the row separated by pipes " | " and spaces.
The table consists of exactly 3 columns "Town", "Latitude" and "Longitude". The latitude and longitude columns always contain valid numbers. 
Output is in the form of objects