Create Movies.js
A function that stores information about movies inside an array. The movies object info is: name, director and date. Input can be of several types:
•	"addMovie {movie name}" – add the movie
•	"{movie name} directedBy {director}" – check if the movie exists and then add the director
•	"{movie name} onDate {date}" – check if the movie exists and then add the date
At the end program prints all the movies that have all the info (if the movie has no director, name or date, don’t print it) in JSON format.