Create StoreProvision.js
Input comes as two arrays. The first array represents a current stock of the local store. The second array contains products which the store has ordered for delivery.
The following information applies to both arrays:
Every even index holds the name of the product and on every odd index holds the quantity of that product. The second array could contain products that are already in the local store. If that happens  the quantity for the given product is to be increased. An object should be used for storage, the final result should be printed as follows : (product -> quantity)