Create MakeADictionary.js
Input comes as an array with strings in the form of JSON's. 
The task is to parse these strings and combine them into one object. Every string from the array holds terms and a description. If the same term is received twice, it should be replaced with the new definition.
Print  format:
Term: ${term} => Definition: ${definition}
Don't forget to sort the dictionary alphabetically by the terms as in real dictionaries.