Create Meetings.js
A function that manages meeting appointments. The input comes as an array of strings. Each string contains a weekday and person’s name. For each successful meeting, a message should be printed. If he same weekday is received twice, the meeting cannot be scheduled so a conflict message should be printed. At the end print result should be a list of all successful meetings.