Create AddressBook.js
A function that stores information about a person’s name and his address. The input comes as an array of strings. Each string contains the name and the address separated by a colon. If a name is received more than once, the address should be replaced. At the end the full list should be printed, sorted alphabetically by the person’s name.