Create Neighbourhoods.js
A function that receives list of neighborhoods and then some people, who are going to live in it. The input comes as array of strings. The first element is the list of neighborhoods separated by ", ". The rest of the elements are in the format: a neighborhood followed by a name of a person - "{neighborhood} - {person}". The task is to add the person to the neighborhood only if the neighborhood is in the list of neighborhoods. At the end the result should be printed as follows:  the neighborhoods sorted by the count of inhabitants in descending order. : 
"{neighborhood}: {inhabitants count}
--{1st inhabitant}
--{2nd inhabitant}