Create WordTracker.js
A function that receives an array of words and finds occurrences of given words in that sentence. 
The input comes as array of strings. The first string contains the words to look for separated by a space. All strings after that are just words. 
For each word the result should show how many times it occurs. The words should be sorted by count in descending.