Create PartyTime.js
Task is to create a function that tackles a list of party guests. Many guests are invited and they are two types: VIP and regular. When guests come to the party, code should check if he/she exists in any of the two reservation lists. 
The input comes as array of strings. The list with the guests given before a command "PARTY" is received. All VIP numbers start with digit.
When command "PARTY" is received, guests start coming.
Result format: Print the count of guests then all guests, who didn't come to the party (VIP must be first)