Create CompanyUsers.js
A function which keeps information about companies and their employees. 
Input is an array of strings containing company name and employee's id. The task is to add each employee to the given company, keeping in mind that a company cannot have two employees with the same id.
The companies should be ordered by the name in ascending order. 
Print the company name and each employee's id in the following format:
{companyName}
-- {id1}
-- {id2}
-- {idN}