Create StringSubstring.js
The input is given as two separated strings.
The function should check given text for containing a given word. The comparison should be case insensitive. Once a match is found,  the word should be printed and the program should stop. 
If the word isnt found, the result should be  "{word} not found!"