Create HardWord.js
Input is an array which holds string and another array. 
The string is a letter from young boy who does not yet know some words. The letter has few holes, these holes are the words unknown to the boy and which must be filled with strings from the array received at the second index. 
If a length of the hole is 4 it should be replaced with string with the same length and so on…