Create PasswordGenerator.js
The task here is to concatenate the first two strings and replace all vowels in the concatenated string with symbols from the third string. First vowel must be replaced with the first character from third string, second vowel with the second character from that string and so on. If the third string is less than the vowels count in the newly formed string you need to start over with character on 0 index. When you replace all vowels reverse the new password and print it on the console in a format:
 'Your generated password is {password}'