Create Furniture.js
A function to calculate the total cost of different types of furniture. Input comes in  lines  until  receiving the line "Purchase". For the line to be valid it should be in the following format:
">>{furniture name}<<{price}!{quantity}"
The price can be floating point number or whole number. The names of furniture should be stored as well as  the total price. 
At the end program should print print the each bought furniture on separate line in the format:
"Bought furniture:
{1st name}
{2nd name}
…"
And on the last line print the following: "Total money spend: {spend money}" formatted to the second decimal point.