Create Race.js
A function that processes information about a race. On the first line will be given a list of participants separated by ", ". On the next few lines until you receiving a line "end of race"will be given some info which will be some alphanumeric characters. In between them there could be some extra characters which should be ignored. 
For example: "G!32e%o7r#32g$235@!2e". The letters are the name of the person and the sum of the digits is the distance he ran. So here we have George who ran 29 km. Task: Store the information about the person only if the list of racers contains the name of the person. If the same person is received more than once just add the distance to his old distance. At the end print the top 3 racers ordered by distance in descending in the format:
"1st place: {first racer}
2nd place: {second racer}
3rd place: {third racer}"