Create ComputerStoreProblem.js
Task Description:
Write a program that prints you a receipt for your new computer. You will receive the prices (without tax) of the parts until you receive what type of customer is this - special or regular. Once you receive the type of the customer you should print the receipt.
The taxes are 20% of each part's price you receive. 
If the customer is special, then he has a 10% discount of the price of the total price with taxes.
If a given price is not positive number, you should print "Invalid price!" on the console and continue with the next price.
If total price is equal to zero, you should print "Invalid order!" on the console.
Input
•	You will receive numbers representing prices (without tax) until command "special" or "regular":
Output
•	The receipt should be in the following format: 

"Congratulations you've just bought a new computer!
Price without taxes: {total price without taxes}$
Taxes: {total amount of taxes}$
-----------
Total price: {total price with taxes}$"