Create CircleArea.js
A function that takes a single argument as an input. 
The task is to check the type of the input argument. 
If it is a number, it should be assumed is the radius of a circle and calculate the circle area. 
If the argument type is NOT a number,  the following text should be printed on the console: 
'We can not calculate the circle area, because we receive a {type of argument}.'
The input comes as a single argument passed to your function.