Create MathOperations.js
A JS function that takes two numbers and a string as an input. The string may be one of the following: '+', '-', '*', '/', '%', '**'.  
The result of the mathematical operation between both numbers and the operator should be printed on the console.
The input comes as two numbers and a string argument passed to your function.
The output should be printed on the console.