Create Fruit.js
A function that calculates how much money you need to buy fruit. You will receive a string for the type of fruit you want to buy, a number for weight in grams and another number for the price per kilogram. 
Print the following text on the console:  
'I need ${money} to buy {weight} kilograms {fruit}.'