Create TimeToWalk.js
A function that calculates how long it takes a student to get to university. 
The function takes three numbers:
•	The first is the number of steps the student takes from their home to the university
•	Тhe second number is the length of the student's footprint in meters
•	Тhe third number is the student speed in km/h
Every 500 meters the student rests and takes a 1 minute break.
The task is to calculate how long the student walks from home to university and print on the console the result in the following format: 'hours:minutes:seconds'.