Create CookingByNumbers.js
A program that receives 6 parameters which are a number and a list of five operations. Operations are to be performed sequentially by starting with the input number and using the result of every operation as starting point for the next one. Result should be printed for every operation in order. The operations can be one of the following:
•	chop - divide the number by two
•	dice - square root of number
•	spice - add 1 to number
•	bake - multiply number by 3
•	fillet - subtract 20% from number