Create PieceOfPie.js
A function that receives three parameters – an array of pie flavors as strings and then two target flavors as strings. The result of the function is a new array, containing a section of the original array, starting at the first target flavor parameter, and ending at (and including) the second target flavor parameter.