Create search.js
An HTML page holds a list of towns, a search box and a [Search] button. Implement the search function to bold and underline the items from the list which include the text from the search box. Also print the amount of items the current search matches in the format "<matches> matches found"