Create solution.js
You will be given some furniture as an array of objects. Each object will have a name, a price and a decoration factor. 
When the "Generate" button is clicked, add a new row to the table for each piece of furniture with image, name, price and decoration factor (code example below). 
When the "Buy" button is clicked, get all checkboxes that are marked and show in the result textbox the names of the piece of furniture that were checked, separated by a comma and single space (", ") in the following format: "Bought furniture: {furniture1} {furniture2}…".
On the next line, print the total price in format: "Total price: {totalPrice}" (formatted to the second decimal point). Finally, print the average decoration factor in the format: "Average decoration factor: {decFactor}"